requirements.txt
feedparser
httpx
PyYAML
python-dotenv
fastapi
uvicorn
